Close OSS device on stream suspend.
Without this, virtual_bt_speaker does not detect that the OSS device is
unplugged if the stream is suspended. That prevents kernel from properly
cleaning up the unplugged device.

A solution with polling on device disconnect would be cleaner, but that does not
work: there is no poll event when OSS device is unplugged.