virtual_oss: Implement SNDCTL_DSP_LOW_WATER ioctl.
This allows to set the low water level for poll() / select() in bytes.
It determines what is considered enough data in the recording buffer for
the application to proceed, or enough free space in the playback buffer
respectively.

The low water level is reset to the default of one fragment when the
buffer size changes. Unlike FreeBSD OSS there is a range check on the
given value. The OSSv4 documentation hints at returning the effective
value here, but FreeBSD soundcard.h defines this ioctl as write only.